added outline, updated component naming, moved avatar props
