+++ src/components/Avatar/Avatar.tsx
@@ -1,6 +1,5 @@
-import { Avatar } from "radix-ui";
-import { AvatarProps } from "./AvatarProps";
-import { avatarVariants } from "./variants";
+import { Avatar as RadixAvatar } from "radix-ui";
+import { avatarVariants, AvatarProps } from "./variants";
 
 /**
  * Avatar Component
@@ -8,13 +7,13 @@ import { avatarVariants } from "./variants";
  * @param { AvatarProps } props
  * @returns Avatar Component
  */
-export const AvatarCustom = (props: AvatarProps) => {
+export const Avatar = (props: AvatarProps) => {
     return (
-    <Avatar.Root>
-		<Avatar.Image className={avatarVariants(props)} src={props.src} alt={props.alt ?? "an avatar"}/>
-		<Avatar.Fallback className={avatarVariants(props)}>{props.fallback.length == 0 ? "NA" : 
+    <RadixAvatar.Root>
+		<RadixAvatar.Image className={avatarVariants(props)} src={props.src} alt={props.alt ?? "an avatar"}/>
+		<RadixAvatar.Fallback className={avatarVariants(props)}>{props.fallback.length == 0 ? "NA" : 
         props.fallback.length <= 2 ? 
-        props.fallback : props.fallback.substring(0, 2)}</Avatar.Fallback>
-	</Avatar.Root>
+        props.fallback : props.fallback.substring(0, 2)}</RadixAvatar.Fallback>
+	</RadixAvatar.Root>
     );
 };
